Simplificación de la page CaseList
