+++ src/App.js
@@ -5,9 +5,8 @@ import MainPage from './pages/MainPage/MaingPage'
 import NewAgent from './pages/NewAgent/NewAgent'
 import NewCase from './pages/NewCase/NewCase'
 import Login from './pages/Login/Login'
-import CaseList from './pages/CaseList/CaseList'
 import CaseDetail from './pages/CaseDetail/CaseDetail'
-import Search from './pages/Search/Search'
+import CaseList from './pages/CaseList/CaseList'
 import ExaDetail from './pages/ExaDetail/ExaDetail'
 
 function App () {
@@ -20,9 +19,8 @@ function App () {
           <Route path='/inicio' element={<MainPage />} />
           <Route path='/nuevo-asesor' element={<NewAgent />} />
           <Route path='/nuevo-caso' element={<NewCase />} />
-          <Route path='/listado-casos' element={<CaseList />} />
           <Route path='/monitoreo/:id' element={<CaseDetail />} />
-          <Route path='/busqueda-avanzada' element={<Search />} />
+          <Route path='/listado-casos' element={<CaseList />} />
           <Route path='/asesor/:exa' element={<ExaDetail />} />
           <Route path='*' element={<h1>404, página no encontrada!</h1>} />
         </Routes>
