+++ src/index.scss
@@ -17,7 +17,5 @@
 @use './pages/Login/login';
 @use './pages/NewAgent/newAgent';
 @use './pages/NewCase/newCase';
-@use './pages/CaseList/caseList';
 @use './pages/CaseDetail/caseDetail';
-@use './pages/Search/search';
 @use './pages/ExaDetail/exaDetails';
