--- src/pages/CaseList/_caseList.scss
@@ -1,13 +1,0 @@
-.case-container{
-    margin: 2em;
-    padding: .5em;
-    box-shadow: 0 0 10px 0 rgba(0,0,0,0.2);
-}
-
-.case-list__form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    margin: 3em;
-}
